+++ test/lightning/ai_assistant/ai_assistant_test.exs
@@ -57,11 +57,26 @@ defmodule Lightning.AiAssistantTest do
       user: user,
       workflow: %{jobs: [job_1 | _]} = _workflow
     } do
+      job_expression = "fn(state => state);\n"
+      adaptor = "@openfn/language-http@7.0.6"
+      message_content = "what?"
+
       session =
         insert(:chat_session,
           user: user,
           job: job_1,
-          messages: [%{role: :user, content: "what?", user: user}]
+          expression: job_expression,
+          adaptor: adaptor,
+          messages: [
+            %{
+              role: :user,
+              content: message_content,
+              user: user,
+              status: :pending,
+              # needed to avoid flaky sorting
+              inserted_at: DateTime.utc_now() |> DateTime.add(-1)
+            }
+          ]
         )
 
       Mox.stub(Lightning.MockConfig, :apollo, fn key ->
@@ -93,7 +108,10 @@ defmodule Lightning.AiAssistantTest do
         {:ok, %Tesla.Env{status: 200, body: reply}}
       end)
 
-      {:ok, updated_session} = AiAssistant.query(session, "foo")
+      {:ok, updated_session} = AiAssistant.query(session, message_content)
+      dbg(updated_session)
+      assert updated_session.expression == job_expression
+      assert updated_session.adaptor == adaptor
 
       assert Enum.count(updated_session.messages) == Enum.count(reply["history"])